requirements.txt
openai
sounddevice
scipy
gtts
playsound
python-dotenv
